requirements.txt
streamlit
numpy
scikit-learn
lightgbm
google-generativeai
requests
